[Bronze III] Title: 종이접기, Time: 92 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/16504. 종이접기/종이접기.cc
@@ -7,7 +7,7 @@ int main(void) {
   cin >> N;
   long long ans = 0;
   for (int i = 0; i < N * N; i++) {
-    long long num;
+    int num;
     cin >> num;
     ans += num;
   }
